Backend/requirements.txt
asgiref==3.11.1
Django==4.2.28
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==25.0.3
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
